Amended tests.
