+++ tests/test_04_SQL_report_as_image.py
