+++ tests/test_07_Create_edit_Instrument_Query.py
