+++ tests/test_08_Instrument_Query_view.py
