+++ tests/test_10_Instrument_Query_as_image.py
