+++ tests/test_14_Record_Table_as_image.py
