+++ tests/test_18_Accumulation_report_view.py
