+++ tests/test_20_Accumulation_report_as_image.py
