+++ tests/test_27_Accumulation_report_public_api.py
