+++ tests/test_31_System_Query_download.py
@@ -1,0 +1,29 @@
+# Generated from Selenium IDE
+# Test name: t31 System Query download
+import pytest
+import time
+import json
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
+
+class Test_31_System_Query_download:
+  def setup_method(self, method):
+    self.driver = self.selectedBrowser
+    self.vars = {}
+  def teardown_method(self, method):
+    self.driver.quit()
+
+  def test_31_System_Query_download(self):
+    self.driver.get("http://127.0.0.1/")
+    self.driver.find_element(By.LINK_TEXT, "My Projects").click()
+    assert len(self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,'Advanced Reports Test')]")) > 0
+    self.driver.find_element(By.LINK_TEXT, "Advanced Reports Test").click()
+    self.driver.find_element(By.LINK_TEXT, "Advanced Reports").click()
+    self.driver.find_element(By.LINK_TEXT, "System Query").click()
+    self.driver.execute_script("$.get( window.location.href.replace('page=view','page=system_view') + '&download=1', function(data,status,xhr){if(data.indexOf('directory_prefix')>-1&&xhr.getResponseHeader('Content-Type')=='text/csv; charset=utf-8'){$('body').append('<span data-csv-ok=\"1\"></span>')};$('body').append('<span data-csv=\"1\"></span>')} )")
+    WebDriverWait(self.driver, 60).until(expected_conditions.presence_of_element_located((By.CSS_SELECTOR, "span[data-csv]")))
+    self.vars["csvDownloadOK"] = self.driver.execute_script("return $('span[data-csv-ok]').length")
+    assert(self.vars["csvDownloadOK"] == "1")
