+++ tests/test_32_System_Query_as_image.py
