--- tests/test_t30SystemQueryview.py
@@ -1,31 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-class TestT30SystemQueryview():
-  def setup_method(self, method):
-    self.driver = webdriver.Firefox()
-    self.vars = {}
-  
-  def teardown_method(self, method):
-    self.driver.quit()
-  
-  def test_t30SystemQueryview(self):
-    self.driver.get("http://127.0.0.1/")
-    self.driver.find_element(By.LINK_TEXT, "My Projects").click()
-    elements = self.driver.find_elements(By.XPATH, "//*[@id=\"table-proj_table\"][contains(.,\'Advanced Reports Test\')]")
-    assert len(elements) > 0
-    self.driver.find_element(By.LINK_TEXT, "Advanced Reports Test").click()
-    self.driver.find_element(By.LINK_TEXT, "Advanced Reports").click()
-    self.driver.find_element(By.LINK_TEXT, "System Query").click()
-    assert self.driver.find_element(By.CSS_SELECTOR, "#mod-advrep-table thead th[data-colnum=\"0\"]").text == "[e][external_module_id]"
-    assert self.driver.find_element(By.CSS_SELECTOR, "#mod-advrep-table thead th[data-colnum=\"1\"]").text == "[e][key]"
-
